// studyflow-mobile/lib/institute-types.ts
export type Batch = {
  id: string
  institute_id: string
  name: string
  course_name: string
  description: string | null
  start_date: string
  end_date: string | null
  class_level: string | null
  total_seats: number | null
  enrolled_students: number
  fee_amount: number
  schedule_days: string[] | null
  schedule_time: string | null
  teacher_name: string | null
  is_active: boolean
  created_at: string
  updated_at: string
}